Move FavoriteNameToPreferredTLFNameFormatAs
...to tlf.
